src/services/remote/requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
msgpack-python==0.5.6
oauthlib==3.0.2
Pyro4==4.76
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
serpent==1.28
urllib3==1.25.3
